+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 
 ## [Unreleased]
 
+### Changed
+- bsmtp: fix and update code change cli parsing to cli11 [PR #1376]
+
 ## [22.0.2] - 2023-02-07
 
 ### Changed
@@ -450,4 +453,5 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 [PR #1363]: https://github.com/bareos/bareos/pull/1363
 [PR #1368]: https://github.com/bareos/bareos/pull/1368
 [PR #1369]: https://github.com/bareos/bareos/pull/1369
+[PR #1376]: https://github.com/bareos/bareos/pull/1376
 [unreleased]: https://github.com/bareos/bareos/tree/master